requirements.txt
google-adk==1.6.1
mcp==1.10.1
python-dotenv==1.1.1